Fixed bug in PrivRemoveInsert pass.
Fixed a bug in which PrivRemoveInsert was adding an additional
privilege to remove in calls to priv_remove().  Fixing the bug
required adding a cast to a bit-shift operation.

+++ lib/Transforms/PrivAnalysis/PrivRemoveInsert.cpp
@@ -71,7 +71,7 @@ void PrivRemoveInsert::addToArgs(std::vector<Value *>& Args,
     int cap = 0;
 
     for (cap = 0; cap < CAP_TOTALNUM; ++cap) {
-        if ((CAPArray & (1 << cap)) == 0) {
+        if ((CAPArray & ((uint64_t)1u << cap)) == 0) {
             continue;
         }
